getting ready for a test

+++ src/main/java/com/hbm/tileentity/machine/TileEntityMachineFENSU.java
@@ -1,7 +1,6 @@
 package com.hbm.tileentity.machine;
 
 import com.hbm.lib.Library;
-
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.math.AxisAlignedBB;
@@ -10,71 +9,68 @@ import net.minecraftforge.fml.relauncher.SideOnly;
 
 public class TileEntityMachineFENSU extends TileEntityMachineBattery {
 
-	public float prevRotation = 0F;
-	public float rotation = 0F;
+    public float prevRotation = 0F;
+    public float rotation = 0F;
 
-	@Override
-	public void update() {
+    @Override
+    public void update() {
 
-		this.maxPower = Long.MAX_VALUE;
+        this.maxPower = Long.MAX_VALUE;
 
-		if(!world.isRemote) {
+        if (!world.isRemote) {
 
-			short mode = (short) this.getRelevantMode();
+            short mode = this.getRelevantMode();
 
-			if(mode == 1 || mode == 2)
-			{
-				age++;
-				if(age >= 20)
-				{
-					age = 0;
-				}
+            if (mode == 1 || mode == 2) {
+                age++;
+                if (age >= 20) {
+                    age = 0;
+                }
 
-				if(age == 9 || age == 19)
-					ffgeuaInit();
-			}
+                if (age == 9 || age == 19)
+                    ffgeuaInit();
+            }
 
-			power = Library.chargeTEFromItems(inventory, 0, power, maxPower);
-			power = Library.chargeItemsFromTE(inventory, 1, power, maxPower);
+            power = Library.chargeTEFromItems(inventory, 0, power, maxPower);
+            power = Library.chargeItemsFromTE(inventory, 1, power, maxPower);
 
-			NBTTagCompound nbt = new NBTTagCompound();
-			nbt.setLong("power", power);
-			nbt.setLong("maxPower", maxPower);
-			nbt.setShort("redLow", redLow);
-			nbt.setShort("redHigh", redHigh);
-			this.networkPack(nbt, 250);
-		} else {
-			this.prevRotation = this.rotation;
-			this.rotation += this.getSpeed();
+            NBTTagCompound nbt = new NBTTagCompound();
+            nbt.setLong("power", power);
+            nbt.setLong("maxPower", maxPower);
+            nbt.setShort("redLow", redLow);
+            nbt.setShort("redHigh", redHigh);
+            this.networkPack(nbt, 250);
+        } else {
+            this.prevRotation = this.rotation;
+            this.rotation += this.getSpeed();
 
-			if(rotation >= 360) {
-				rotation -= 360;
-				prevRotation -= 360;
-			}
-		}
-	}
+            if (rotation >= 360) {
+                rotation -= 360;
+                prevRotation -= 360;
+            }
+        }
+    }
 
-	@Override
-	public long getPowerRemainingScaled(long i) {
+    @Override
+    public long getPowerRemainingScaled(long i) {
 
-		double powerScaled = (double)power / (double)maxPower;
+        double powerScaled = (double) power / (double) maxPower;
 
-		return (long)(i * powerScaled);
-	}
+        return (long) (i * powerScaled);
+    }
 
-	public float getSpeed() {
-		return (float) Math.pow(Math.log(power * 0.75 + 1) * 0.05F, 5);
-	}
+    public float getSpeed() {
+        return (float) Math.pow(Math.log(power * 0.75 + 1) * 0.05F, 5);
+    }
 
-	@Override
-	public AxisAlignedBB getRenderBoundingBox() {
-		return TileEntity.INFINITE_EXTENT_AABB;
-	}
+    @Override
+    public AxisAlignedBB getRenderBoundingBox() {
+        return TileEntity.INFINITE_EXTENT_AABB;
+    }
 
-	@Override
-	@SideOnly(Side.CLIENT)
-	public double getMaxRenderDistanceSquared()
-	{
-		return 65536.0D;
-	}
+    @Override
+    @SideOnly(Side.CLIENT)
+    public double getMaxRenderDistanceSquared() {
+        return 65536.0D;
+    }
 }
