+++ src/main/java/glmath/glm/vec/_2/d/Vec2d.java
@@ -6,7 +6,6 @@
 package glmath.glm.vec._2.d;
 
 /**
- *
  * @author elect
  */
 public class Vec2d extends FuncRelational {
@@ -17,7 +16,7 @@ public class Vec2d extends FuncRelational {
     }
 
     public Vec2d(int x, int y) {
-        this((double) x, (double) y);
+        this(x, (double) y);
     }
 
     public Vec2d(double x, double y) {
